backend/requirements.txt
flask==3.0.0
google-auth==2.25.2
google-api-python-client==2.110.0
python-dotenv==1.0.0
